requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.4
Django==2.2.6
django-debug-toolbar==3.2
django-extensions==3.1.0
djangorestframework==3.12.2
flake8==3.8.4
idna==2.8
importlib-metadata==1.5.0
isort==5.7.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==8.1.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
urllib3==1.25.6
wcwidth==0.1.8
whitenoise==5.2.0
zipp==2.2.0
